+++ CHANGES.md
@@ -3,8 +3,10 @@
 Released: -
 
 - Add 'docs_oauth2_redirect_path_external' parameter to support absolute redirect url ([issue #602][issue_602]).
+- Update the redoc's default settings to use jsDelivr by default instead ([pr #650](pr_650)).
 
 [issue_602]: https://github.com/apiflask/apiflask/issues/602
+[pr_650]: https://github.com/apiflask/apiflask/pull/650
 
 
 ## Version 2.3.2